move calendar service to services
